Fix: ピッチ読み込み値がおかしくなる不具合 Fix: 調整用SliderがThanksPanelより手前に表示される不具合

+++ Assets/Scripts/SoundTester.cs
@@ -55,7 +55,7 @@ public class SoundTester : MonoBehaviour
 
         idText.text = nowPlaying.ToString("D3");
         pitchText.text = audioSources[nowPlaying].pitch.ToString("0.000");
-        s.value = (int)((audioSources[nowPlaying].pitch - 0.75f) / 0.025f);
+        s.value = ((audioSources[nowPlaying].pitch - 0.75f) / 0.025f);
     }
 
     public void UpdatePitchText()
